refactor: helpers to parse benchmark arguments
Remove code duplication between the Storage and Pub/Sub benchmarks. Even
I don't want to cut&paste this code one more time. Did some minor
simplifications, like using `absl::DurationParse` instead of a custom
parser.
