+++ google/cloud/storage/benchmarks/CMakeLists.txt
@@ -97,12 +97,8 @@ if (BUILD_TESTING)
     # List the unit tests, then setup the targets and dependencies.
     set(storage_benchmarks_unit_tests
         # cmake-format: sort
-        benchmark_make_random_test.cc
-        benchmark_parse_args_test.cc
-        benchmark_parser_test.cc
-        benchmark_utils_test.cc
-        throughput_options_test.cc
-        throughput_result_test.cc)
+        benchmark_make_random_test.cc benchmark_parser_test.cc
+        throughput_options_test.cc throughput_result_test.cc)
 
     foreach (fname ${storage_benchmarks_unit_tests})
         google_cloud_cpp_add_executable(target "storage_benchmarks" "${fname}")
