+++ google/cloud/storage/benchmarks/benchmark_parser_test.cc
@@ -50,7 +50,6 @@ TEST(StorageBenchmarksUtilsTest, ParseDuration) {
   using s = std::chrono::seconds;
   using m = std::chrono::minutes;
   using h = std::chrono::hours;
-  EXPECT_EQ(s(42).count(), ParseDuration("42").count());
   EXPECT_EQ(s(m(42)).count(), ParseDuration("42m").count());
   EXPECT_EQ(s(h(3)).count(), ParseDuration("3h").count());
   EXPECT_EQ(s(1800).count(), ParseDuration("1800s").count());
