+++ google/cloud/storage/benchmarks/benchmark_utils.cc
@@ -25,139 +25,6 @@
 namespace google {
 namespace cloud {
 namespace storage_benchmarks {
-bool EndsWith(std::string const& val, std::string const& suffix) {
-  auto pos = val.rfind(suffix);
-  if (pos == std::string::npos) {
-    return false;
-  }
-  return pos == val.size() - suffix.size();
-}
-
-// This parser does not validate the input fully, but it is good enough for our
-// purposes.
-std::int64_t ParseSize(std::string const& val) {
-  long s = std::stol(val);  // NOLINT(google-runtime-int)
-  if (EndsWith(val, "TiB")) {
-    return s * kTiB;
-  }
-  if (EndsWith(val, "GiB")) {
-    return s * kGiB;
-  }
-  if (EndsWith(val, "MiB")) {
-    return s * kMiB;
-  }
-  if (EndsWith(val, "KiB")) {
-    return s * kKiB;
-  }
-  if (EndsWith(val, "TB")) {
-    return s * kTB;
-  }
-  if (EndsWith(val, "GB")) {
-    return s * kGB;
-  }
-  if (EndsWith(val, "MB")) {
-    return s * kMB;
-  }
-  if (EndsWith(val, "KB")) {
-    return s * kKB;
-  }
-  return s;
-}
-
-std::size_t ParseBufferSize(std::string const& val) {
-  auto const s = ParseSize(val);
-  if (s < 0 || static_cast<std::uint64_t>(s) >
-                   (std::numeric_limits<std::size_t>::max)()) {
-    internal::ThrowRangeError("invalid range in ParseBufferSize");
-  }
-  return static_cast<std::size_t>(s);
-}
-
-// This parser does not validate the input fully, but it is good enough for our
-// purposes.
-std::chrono::seconds ParseDuration(std::string const& val) {
-  long s = std::stol(val);  // NOLINT(google-runtime-int)
-  if (EndsWith(val, "h")) {
-    return s * std::chrono::seconds(std::chrono::hours(1));
-  }
-  if (EndsWith(val, "m")) {
-    return s * std::chrono::seconds(std::chrono::minutes(1));
-  }
-  if (EndsWith(val, "s")) {
-    return s * std::chrono::seconds(std::chrono::seconds(1));
-  }
-  return std::chrono::seconds(s);
-}
-
-absl::optional<bool> ParseBoolean(std::string const& val) {
-  if (val.empty()) {
-    return {};
-  }
-  auto lower = val;
-  std::transform(lower.begin(), lower.end(), lower.begin(),
-                 [](char x) { return static_cast<char>(std::tolower(x)); });
-  if (lower == "true") return true;
-  if (lower == "false") return false;
-  return {};
-}
-
-std::string Basename(std::string const& path) {
-  // With C++17 we would use `std::filesystem::path`, until then do the poor's
-  // person version.
-#if _WIN32
-  return path.substr(path.find_last_of("\\/") + 1);
-#else
-  return path.substr(path.find_last_of('/') + 1);
-#endif  // _WIN32
-}
-
-std::string BuildUsage(std::vector<OptionDescriptor> const& desc,
-                       std::string const& command_path) {
-  std::ostringstream os;
-  os << "Usage: " << Basename(command_path) << " [options] <region>\n";
-  for (auto const& d : desc) {
-    os << "    " << d.option << ": " << d.help << "\n";
-  }
-  return std::move(os).str();
-}
-
-std::vector<std::string> OptionsParse(std::vector<OptionDescriptor> const& desc,
-                                      std::vector<std::string> argv) {
-  if (argv.empty()) {
-    return argv;
-  }
-  std::string const usage = BuildUsage(desc, argv[0]);
-
-  auto next_arg = argv.begin() + 1;
-  while (next_arg != argv.end()) {
-    std::string const& argument = *next_arg;
-
-    // Try to match `argument` against the options in `desc`
-    bool matched = false;
-    for (auto const& d : desc) {
-      if (argument.rfind(d.option, 0) != 0) {
-        // Not a match, keep searching
-        continue;
-      }
-      std::string val = argument.substr(d.option.size());
-      if (!val.empty() && val[0] != '=') {
-        // Matched a prefix of an option, keep searching.
-        continue;
-      }
-      if (!val.empty()) {
-        // The first character must be '=', remove it too.
-        val.erase(val.begin());
-      }
-      d.parser(val);
-      // This is a match, consume the argument and stop the search.
-      matched = true;
-      break;
-    }
-    // If next_arg is matched against any option erase it, otherwise skip it.
-    next_arg = matched ? argv.erase(next_arg) : next_arg + 1;
-  }
-  return argv;
-}
 
 #if GOOGLE_CLOUD_CPP_HAVE_GETRUSAGE
 namespace {
@@ -237,33 +104,6 @@ bool SimpleTimer::SupportPerThreadUsage() {
 #else
   return false;
 #endif  // GOOGLE_CLOUD_CPP_HAVE_RUSAGE_THREAD
-}
-
-std::string FormatSize(std::uintmax_t size) {
-  struct {
-    std::uintmax_t limit;
-    std::uintmax_t resolution;
-    char const* name;
-  } ranges[] = {
-      {kKiB, 1, "B"},
-      {kMiB, kKiB, "KiB"},
-      {kGiB, kMiB, "MiB"},
-      {kTiB, kGiB, "GiB"},
-  };
-  std::uintmax_t resolution = kTiB;
-  char const* name = "TiB";
-  for (auto const& r : ranges) {
-    if (size < r.limit) {
-      resolution = r.resolution;
-      name = r.name;
-      break;
-    }
-  }
-  std::ostringstream os;
-  os.setf(std::ios::fixed);
-  os.precision(1);
-  os << (static_cast<double>(size) / static_cast<double>(resolution)) << name;
-  return os.str();
 }
 
 void DeleteAllObjects(google::cloud::storage::Client client,
