+++ google/cloud/storage/benchmarks/storage_benchmarks_unit_tests.bzl
@@ -18,9 +18,7 @@
 
 storage_benchmarks_unit_tests = [
     "benchmark_make_random_test.cc",
-    "benchmark_parse_args_test.cc",
     "benchmark_parser_test.cc",
-    "benchmark_utils_test.cc",
     "throughput_options_test.cc",
     "throughput_result_test.cc",
 ]
