+++ google/cloud/storage/benchmarks/storage_parallel_uploads_benchmark.cc
@@ -302,12 +302,14 @@ int main(int argc, char* argv[]) {
 
 namespace {
 
+using ::google::cloud::testing_util::OptionDescriptor;
+
 google::cloud::StatusOr<Options> ParseArgsDefault(
     std::vector<std::string> argv) {
   Options options;
   bool wants_help = false;
   bool wants_description = false;
-  std::vector<gcs_bm::OptionDescriptor> desc{
+  std::vector<OptionDescriptor> desc{
       {"--help", "print usage information",
        [&wants_help](std::string const&) { wants_help = true; }},
       {"--description", "print benchmark description",
@@ -358,9 +360,9 @@ google::cloud::StatusOr<Options> ParseArgsDefault(
          options.maximum_sample_count = std::stol(val);
        }},
   };
-  auto usage = gcs_bm::BuildUsage(desc, argv[0]);
+  auto usage = BuildUsage(desc, argv[0]);
 
-  auto unparsed = gcs_bm::OptionsParse(desc, argv);
+  auto unparsed = OptionsParse(desc, argv);
   if (wants_help) {
     std::cout << usage << "\n";
   }
