+++ google/cloud/testing_util/google_cloud_cpp_testing.bzl
@@ -21,6 +21,7 @@ google_cloud_cpp_testing_hdrs = [
     "capture_log_lines_backend.h",
     "check_predicate_becomes_false.h",
     "chrono_literals.h",
+    "command_line_parsing.h",
     "contains_once.h",
     "crash_handler.h",
     "example_driver.h",
@@ -35,6 +36,7 @@ google_cloud_cpp_testing_hdrs = [
 google_cloud_cpp_testing_srcs = [
     "assert_ok.cc",
     "capture_log_lines_backend.cc",
+    "command_line_parsing.cc",
     "crash_handler.cc",
     "example_driver.cc",
     "scoped_environment.cc",
