+++ google/cloud/testing_util/google_cloud_cpp_testing_unit_tests.bzl
@@ -18,6 +18,7 @@
 
 google_cloud_cpp_testing_unit_tests = [
     "assert_ok_test.cc",
+    "command_line_parsing_test.cc",
     "contains_once_test.cc",
     "crash_handler_test.cc",
     "example_driver_test.cc",
